refactor: reorganize active session monitoring utilities
